Sample Criteria Extension for Multinomial Regression

Calculate sample criteria extended for multinominal regression.

Usage

pmsamplesize(
  Q,
  k,
  p,
  r2_cs = NULL,
  auc = NULL,
  prev,
  shrinkage = 0.9,
  r2_nagelkerke = 0.15,
  sigma = 0.05,
  margin_error = 0.05
)

Arguments

Q

An integer. The number of candidate predictors. See Details

k

An integer. The number of levels in response variable

p

A numeric vector. Outcome proportion(s)

r2_cs

numeric. Cox-Snell \(R^2\).

auc

numeric. Pairwise C statistics

prev

numeric. Outcome proportions in categories relative to reference

shrinkage

A pre-defined shrinkage factor. Default to 0.9

r2_nagelkerke

An adjustment for Cox-Snell \(R^2\). Default to 0.15.

sigma

A double. tolerate rate between adjusted \(R^2\) and apparent \(R^2\). Default to 0.05

margin_error

A divergence for margin error specified by Chi-sqaure distribution's confidence level. Deault to 0.05

Value

An integer

Details

Sample Criteria for Clinical Prediction Model are based on the work of Riley 2018 and Calster 2020

References

Examples

K <- 5
events <- c(2557, 186, 176, 467, 120)
c_stats <- c(0.85, 0.92, 0.99, 0.95, 0.75, 0.95, 0.87, 0.87, 0.71, 0.82)
prev <- phi_pairs(events)
set.seed(1234)
pmsamplesize(Q = 30, k = 2, auc = 0.81, prev = 0.77)
#> NB: Assuming 0.05 acceptable difference in apparent & adjusted R-squared
#> NB: Assuming 0.05 margin of error in estimation of intercept
#> NB: Events per Predictor Parameter assumes prevalance 0.77 
#>            sample_size shrinkage parameter    CS_Rsq   Max_Rsq   Nag_Rsq
#> Criteria 1        1129 0.9000000        30 0.2100957 0.3183689 0.3183689
#> Criteria 2        1674 0.9295688        30 0.2100957 0.3183689 0.3183689
#> Criteria 3         273 0.9295688        30 0.2100957 0.3183689 0.3183689
#> Final             1674 0.9295688        30 0.2100957 0.3183689 0.3183689
#>                 EPP
#> Criteria 1 28.97767
#> Criteria 2 42.96600
#> Criteria 3  7.00700
#> Final      42.96600
#> Minimum sample size required for new model development based on user inputs = 1674,
#>  with 1289 events (assuming an outcome prevalence = 0.77) and an EPP = 43