Sample Criteria Extension for Multinomial Regression

Calculate sample criteria extended for multinominal regression.

Usage

pmsamplesize(
  Q,
  k,
  p,
  adjust = FALSE,
  r2_cs_app = NULL,
  r2_cs_adj = NULL,
  shrinkage = 0.9,
  r2_nagelkerke = 0.15,
  auc = NULL,
  prev = NULL,
  sigma = 0.05,
  margin_error = 0.05
)

Arguments

Q

An integer. The number of candidate predictors. See Details

k

An integer. The number of levels in response variable

p

A numeric vector. Outcome proportion(s) or event observations

adjust

A logical. Should adjusted R-square value be used? Default to FALSE

r2_cs_app

A numeric vector. The apparent Cox-Snell \(R^2\)

r2_cs_adj

A numeric vector. The optimism-adjusted Cox-Snell \(R^2\)

shrinkage

A pre-defined shrinkage factor. Default to 0.9

r2_nagelkerke

An adjustment for Cox-Snell \(R^2\). Default to 0.15.

auc

A numeric vector. Pairwise C statistics

prev

A numeric vector. Outcome proportions in categories relative to reference

sigma

A double. Tolerate rate between adjusted \(R^2\) and apparent \(R^2\). Default to 0.05

margin_error

A double. A divergence for margin error specified by Chi-square distribution's confidence level. Default to 0.05

Value

A pmsample class object

final: A single row data frame representing the selected criterion for the minimum sample size for developing a new model

criteria: A data frame containing all the sample sizes based on three criteria. Divergences in sample sizes due to number rounding

sigma: A single numeric representing the tolerate rate between adjusted \(R^2\) and apparent \(R^2\)

margin_error: A divergence for margin error specified by Chi-square distribution's confidence level

Details

The pmsample objects created by pmsamplesize can represent the minimum sample size needed for developing a new model based on the current studies.

pmsamplesize function adapts the amount of information based on the type of model. A certain parameters need pre-specifying for calculation:

The number of candidate predicators Q, usually obtained using the fractional polynomial method and it is not addressed by this package.

The number of levels in response variable k, typically for a binary model k = 2, for multinomial models k > 2. One of the argument between k or p, the outcome proportions or the observation counts of the events, must be specified. In the case of a missing k, it will then defined by the length of p.

The order of p matters for calculation: for a binary model, the second value in the vector is perceived as 'outcome proportion' or 'outcome events'.

Pairwise C-statistics auc for simulation. The process requires estimates of the pairwise outcome proportions \(\phi_{k,r}\) of category k relative to the reference category r, prev. If not supplied, they will be derived from p.

References

Examples

# when only C statistics is reported in a binary model
set.seed(1234)
pmsamplesize(Q = 30, k = 2, r2_nagelkerke = 0.15, auc = 0.81, prev = 0.77)
#> Warning: Inadequate information to calculate sample size for criterion 2
#> NB: Assuming 0.05 acceptable difference in apparent & adjusted R-squared
#> NB: Assuming 0.05 margin of error in estimation of intercept
#> NB: Events per Predictor Parameter assumes prevalance 0.77 
#>            sample_size shrinkage parameter CS_Rsq Max_Rsq Nag_Rsq   EPP
#> Criteria 1        1130       0.9        30   0.21      NA    0.15 29.00
#> Criteria 3         273       0.9        30   0.21      NA    0.15  7.01
#> Final             1130       0.9        30   0.21      NA    0.15 29.00
#> Minimum sample size required for new model development based on user inputs = 1130,
#>  with 870 events (assuming an outcome prevalence = 0.77) and an EPP = 29